Changed feedforward_torque resolution in InitializeCommand() to kInt32

+++ src/mjbots_hardware_interface.cpp
@@ -25,7 +25,7 @@ void MjbotsHardwareInterface::InitializeCommand() {
   }else{
     res.position = ::mjbots::moteus::Resolution::kIgnore;
     res.velocity = ::mjbots::moteus::Resolution::kIgnore;
-    res.feedforward_torque = ::mjbots::moteus::Resolution::kInt16;
+    res.feedforward_torque = ::mjbots::moteus::Resolution::kInt32;
     res.kp_scale = ::mjbots::moteus::Resolution::kIgnore;
     res.kd_scale = ::mjbots::moteus::Resolution::kIgnore;
     res.maximum_torque = ::mjbots::moteus::Resolution::kIgnore;
